Move coreTSModule to client-modules folder
